move publish methods to a separate file
